Make sure existing adapters compatible with reasoning models
